Abstraction edits
Made changes to lookup and declared. When declaring a variable without assigning a value, adds an atom 'nul to the list of values in the state.

Three new abstractions: firstVar, the first variable in the state, firstVal, the value of the first variable, and listVars, a list of all the declared variables